feat: A complete rewrite from scratch!
So my first pass at this file system left several things to be
desired, so much that we stopped using the library pretty quick.

This is a rewrite taking into many of those lessons as well as some
things I've learned recently.

The most important thing to notice here is that this project depends
on an FTL like Dhara to work. It assumes that we can rewrite sectors
on disk, but does it's best to minimize the number of writes that are
done. It is also optimized for specific use cases, and that is mostly
those that you'd find in sensor logging systems.

For example, you can create multiple files and append to
them. Only appends are supported. It's got a very useful/flexible
attributes system inspired by littlefs's that can be used to
store/track metadata alongside files for indexing/application
purposes.

There's still quite a lot of work to do here but this represents a
good starting point to work from.