fix: Read delimiters across sectors when skipping records.
This fixes a bug in how varint_decoder was being used to skip records,
which would error instead of continuing to read a multi-byte delimiter
that spans a sector boundary.

The new behavior is to continue reading a delimiter until done and return
the numbers of bytes cumulatively read.

Initially cropped up as failing resumed downloads that weren't seeking
properly due to the skipping bug.